requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.46
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.20
pydantic-settings==2.7.1
python-dotenv==1.0.1
email-validator==2.3.0
psycopg2-binary==2.9.10
slowapi==0.1.9
cryptography==44.0.0
anthropic==0.43.0
openai==1.59.0
resend==2.6.0
supabase==2.11.0
